refactor: rename template files to examples, fix slug hyphen collapsing
Address review feedback:
- Rename all files: agent_templates → agent_examples, template-card → example-card
- Fix slug_from_name to collapse consecutive hyphens
