+++ apps/ui/src/components/agents/example-card.tsx
