+++ apps/ui/src/components/agents/index.ts
@@ -1,2 +1,2 @@
 export { AgentCard } from "./agent-card";
-export { ExampleCard } from "./template-card";
+export { ExampleCard } from "./example-card";
