+++ crates/server/src/app_builder.rs
@@ -511,7 +511,7 @@ impl ServerAppBuilder {
         );
         let commands_state =
             api::commands::AppState::new(capability_service.clone(), auth_state.clone());
-        let agent_templates_state = api::agent_templates::AppState {
+        let agent_examples_state = api::agent_examples::AppState {
             agent_service: Arc::new(services::AgentService::new(db.clone())),
             capability_service: capability_service.clone(),
             auth: auth_state.clone(),
@@ -614,7 +614,7 @@ impl ServerAppBuilder {
         // Phase 6: Build API router
         // =====================================================================
         let mut api_routes = Router::new()
-            .merge(api::agent_templates::routes(agent_templates_state))
+            .merge(api::agent_examples::routes(agent_examples_state))
             .merge(api::agents::routes(agents_state))
             .merge(api::apps::routes(apps_state))
             .merge(api::harnesses::routes(harnesses_state))
